Track calculation result with global variable
Will help with logic to decide what to do with calculation result when
a new number is entered.

The calculation result was stored in currInput variabe, but entering a
new number would add on to that currInput variable when it should erase
the value and build currInput from scratch.

The only time the result should be kept is when clicking an operator
(+-*/) button, which allows the result to be used in the next
calculation.